+++ Assets/01_Scripts/Base.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44558bfceb6a9c548a09097e89bfd99d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
