+++ Assets/01_Scripts/Base/BaseStatusData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8d913e4e8b005e48b2ab0fd2d2df7ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
