+++ Assets/01_Scripts/Base/BaseStatusSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a6f0ee68b0ac2b41b04e2e7956cef2e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
