+++ Assets/01_Scripts/Unit/UnitStatusSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fef3eba8aa245ba48bac3a2c9b60fbf7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
